Initial SSAO term generation
Add R2SSAO* types and a filter, R2FilterSSAO, used to calculate the
raw SSAO term for a given G-buffer.

Update the interfaces of all renderers to take
AreaInclusiveUnsignedLType values instead of
AreaInclusiveUnsignedIType, for uniformity with textures.

Update the example code interface to take AreaInclusiveUnsignedLType
values instead of AreaInclusiveUnsignedIType, for uniformity with
textures, renderers, and filters.

All renderers and filters now configure their viewports prior to
rendering. As a consequence of this, all renderers and filters now
require access to the inclusive area of the framebuffer to which
they are rendering.

Affects #20