Serious SSAO fixes
The Z component of the noise textures should be 0, because the
texture is supposed to represent a rotation around the Z axis.

  io7m-r2-core/src/main/java/com/io7m/r2/core/filters/R2SSAONoiseTexture.java

The range check in the SSAO shader was incorrect: It was, in effect,
stating that a sample position that is _closer_ to the viewer should
be occluded by a surface that is _further away_. This is exactly the
wrong way round and was causing halo artifacts.

  io7m-r2-shaders/src/main/glsl/com/io7m/r2/shaders/sources/R2SSAO.frag

Affects #20