Remove separate volume/origin transforms
This removes most of the changes that were added to provide separate
light volume/origins. This is now provided by stencil-clipped volumes,
so most of the changes are no longer necessary. The shader interface
changes have not been reverted, and the matrix type changes have also
been kept for future expansion.

Affects #43
Affects #41