Switch to JState for shader state handling
This uses the JState library to handle shader states. This has the
advantage of providing all of the valid state transitions in one
location and immediately signalling errors on any invalid state
transition.

This also uncovered a bug in the emission filter: The wrong shader
was being deactivated and the previous state checking implementation
was not strong enough to catch this bug.

Fix #64