Large update to entire package
This is a very large update that covers most of the package. It
renames the project to follow the latest naming conventions and uses
the latest versions of all of the core io7m libraries. In particular,
it uses the new `jregions` and `jtensors` packages that shift the
bias of the package back towards immutable objects for correctness
and to take advantage of JVM escape analysis in newer JVMs.

It also replaces the old "main" interface with a new "facade" package
that provides more convenient ways to construct the various objects
used by the API and also enables the package to published as a service
in both OSGi and via ServiceLoader.

Additionally, it changes various aspects of the API to expose concrete
types in an attempt to address #94, although this is incomplete.

See: http://blog.io7m.com/2017/02/25/java-module-renaming.xhtml

Fix #93
Fix #91
Fix #92