Replace verifiers with abstract shaders
This replaces the shader verification interfaces with a set of abstract
shaders that enforce the calling protocol and other preconditions
and invariants. The abstract implementations reduce the megamorphism
of the call sites made to shader methods by allowing each shader
implementation class to be exposed directly as opposed to being
exposed via interfaces.

Fix #94