Parameterize light shader outputs
This adds the ability to define R2_LIGHT_SHADER_OUTPUT_TARGET to one
of:

  R2_LIGHT_SHADER_OUTPUT_TARGET_LBUFFER 0x1000
  R2_LIGHT_SHADER_OUTPUT_TARGET_IBUFFER 0x1001

The default shading environment is configured to
R2_LIGHT_SHADER_OUTPUT_TARGET_LBUFFER for backwards compatibility. The
light renderer has a new API entry point that controls rendering to
an image instead of a light buffer, but the actual code
is not yet implemented.

Affects #89