Modularize the codebase
This breaks the core module into a lot of separate modules. Shaders
are divided into shader modules, filters are divided into their
own modules, and the dependencies of all modules are cleaned up
and verified.

Additionally, the default is now to compile light shaders for image
buffers instead of light buffers.

Fix #98